auth-service/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
httpx>=0.24.1
pyjwt>=2.6.0
pydantic>=1.10.8
pydantic[email]>=1.10.8
motor>=3.1.1
pymongo>=4.3.3
python-multipart>=0.0.6
